requirements.txt
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
flask==3.0.3
python-dotenv==1.0.1

# Optional (add later when integrating LLM)

openai==1.35.13
fastapi==0.111.0
uvicorn==0.30.1

langchain==0.2.10
langchain-openai==0.1.7
langchain-community==0.2.10
langchain-text-splitters==0.2.2

streamlit==1.35.0
requests==2.32.3

httpx==0.27.0